New joystick control package
