+++ src/Teleop-Control/joystick_control/package.xml
@@ -4,17 +4,11 @@
   <name>joystick_control</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
-  <maintainer email="connor.needham2015@gmail.com">connor</maintainer>
+  <maintainer email="PSPuttaguna@gmail.com">vscode</maintainer>
   <license>TODO: License declaration</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
-  <depend>rclcpp</depend>
-  <depend>std_msgs</depend>
-  <depend>geometry_msgs</depend>
-  <depend>sensor_msgs</depend>
-  <depend>ros_phoenix</depend>
-  <depend>interfaces</depend>
-  <depend>joy</depend>
+
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
